requirements.txt
pyfiglet
colorama
requests
google-generativeai
